requirements.txt
Django==3.2.8
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
